Added PUR functionality
Restored ID functionality hidden behind debug statement (requested on forum)
For places with a single PUR, it now shows an icon for the type of PUR: plus sign for new places, camera for pictures, pencil for updates, and a flag for flagged places. For multiple PURs on a single place, it still shows the number, but lists the PUR types in the tooltip when hovering over the number. The column is also sortable by the type.