Grabbing the issue-tracker sidepanel which always seems to be available at initialization.
In WME v.2.323 an optimization was made that seems to not load all sidebar tabs at startup, “Settings” being one of them. The script tries to get this sidebar via “sidebar-prefs” but unless that “Settings” sidebar is open when UR-MP initializes, it will fail.

Dumping all of the elements at UR-MP init time shows that sidepanel-issue-tracker is always present, so this can be used in place of sidebar-prefs and should never be null.